Created the app
Made the app and did some front end and backend work. Created the headers for the two pages. Added the carousel functionality but did not implement the backend functionality to fetch the things to fill the carousels with yet. The post method needs to be tested to ensure that the thumbnail and image upload work correctly.

Need to add:
  Backend:
-functionality for getting thumbnails for carousels

  Frontend:
-entire upload page
-ensuring carousels work as intended
-better video player
-make everything look nice