Changed and added routes
Changed the format of thumbnails to contain the id of its associated video instead of the other way around. Also added a route to get all thumbnail ids of a specific genre for the carousels, and a route to get the image and video id of a thumbnail for each item in the carousel. This needs to be tested though.